Split off some common test functions into utils files
