+++ openmdao/recorders/tests/recorder_test_utils.py
@@ -1,0 +1,8 @@
+import time
+
+def run_driver(problem):
+    t0 = time.time()
+    problem.run_driver()
+    t1 = time.time()
+    return t0, t1
+
